Move BetsList widget to their own class file
