+++ lib/widgets/BetsList/BetsList.dart
@@ -1,0 +1,32 @@
+import 'package:flutter/material.dart';
+import 'package:long_term_bets/consumers/BetsConsumer.dart';
+import 'package:long_term_bets/data/Bets.dart';
+import 'package:long_term_bets/providers/BetProvider.dart';
+import 'package:long_term_bets/widgets/BetCard/BetCard.dart';
+
+class BetsList extends StatelessWidget with BetsConsumer, BetProvider {
+  BetsList({
+    @required this.betsType,
+  });
+
+  final BetsType betsType;
+
+  Widget _buildCards(BuildContext context) {
+    final List<Bet> betsToShow = betsList(context, betsType);
+    return ListView.builder(
+        padding: const EdgeInsets.all(5.0),
+        itemBuilder: (BuildContext context, int i) {
+          if (i < betsToShow.length) {
+            return provideBet(
+              betsToShow[i],
+              BetCard(betsType: betsType)
+            );
+          }
+        });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return _buildCards(context);
+  }
+}
